Kes/board review

+++ StudentSoup/src/main/frontend/src/components/common/restaurantNavbar.tsx
@@ -11,11 +11,21 @@ const RestaurantNavbar = () => {
   const history = useHistory();
 
   const IMAGE_FILE_ID = String(sessionStorage.getItem('fileName'));
+  const userId = sessionStorage.getItem('id');
 
+  const [isLogin, setIsLogin] = useState<boolean>(false);
   const [searchSchool, setSearchSchool] = useState<any[]>();
   const [posts, setPosts] = useState<any[]>();
   const [schoolValue, setSchoolValue] = useState<string>();
   const [inputSchool, setInputSchool] = useState<string>();
+
+  useEffect(() => {
+    if (userId === null) {
+      setIsLogin(false);
+    } else {
+      setIsLogin(true);
+    }
+  }, []);
 
   const getSchool = () => {
     axios
@@ -49,6 +59,7 @@ const RestaurantNavbar = () => {
   };
 
   const handleClick = (e: any) => {
+    setInputSchool(e.target.innerText);
     const resultArray = posts?.filter(post => post.schoolName.includes(e.target.innerText));
     const reducc = resultArray?.reduce(function (acc, cur) {
       return [...acc, ...cur];
@@ -76,21 +87,34 @@ const RestaurantNavbar = () => {
     }
   };
 
-  const handleClickLogout = () => {
-    if (sessionStorage.getItem('saved') === String(true)) {
-      sessionStorage.removeItem('email');
-      sessionStorage.removeItem('nickname');
-      sessionStorage.removeItem('departmentId');
-      sessionStorage.removeItem('departmentName');
-      sessionStorage.removeItem('fileName');
-      sessionStorage.removeItem('memberId');
-      sessionStorage.removeItem('schoolId');
-      sessionStorage.removeItem('schoolName');
-      sessionStorage.removeItem('registrationDate');
-      history.push('/');
+  const handleLogin = (e: any) => {
+    const value = e.target.innerText;
+    if (value === 'LOGOUT') {
+      if (sessionStorage.getItem('saved') === String(true)) {
+        sessionStorage.removeItem('email');
+        sessionStorage.removeItem('nickname');
+        sessionStorage.removeItem('departmentId');
+        sessionStorage.removeItem('departmentName');
+        sessionStorage.removeItem('fileName');
+        sessionStorage.removeItem('memberId');
+        sessionStorage.removeItem('schoolId');
+        sessionStorage.removeItem('schoolName');
+        sessionStorage.removeItem('registrationDate');
+        history.push('/');
+      } else {
+        sessionStorage.clear();
+        history.push('/');
+      }
+    } else if (value === 'LOGIN') {
+      history.push('/login');
+    }
+  };
+
+  const handleMypage = () => {
+    if (userId === null) {
+      alert('로그인 후 이용가능합니다.');
     } else {
-      sessionStorage.clear();
-      history.push('/');
+      history.push('/mypage');
     }
   };
 
@@ -102,7 +126,7 @@ const RestaurantNavbar = () => {
             history.push('/');
           }}
         />
-        <div className="w-[466px] h-[44px] px-[23px] flex items-center gap-x-3 border-none rounded-[5px] bg-[#E8E8E8]">
+        <div className="w-[466px] h-[44px] pl-[23px] flex items-center gap-x-3 border-none rounded-[5px] bg-[#E8E8E8]">
           <svg
             width="25"
             height="24"
@@ -123,23 +147,27 @@ const RestaurantNavbar = () => {
             placeholder="학교 명을 입력하세요"
             className="w-full text-[#717171] bg-transparent outline-0"
           ></input>
-          <button onClick={handlePushRestaurant} className="hidden">
+          <button
+            onClick={handlePushRestaurant}
+            className="flex items-center justify-center w-[80px] h-full bg-[#FF611D] text-white rounded-[5px]"
+          >
             검색
           </button>
         </div>
-        <div className="absolute top-[70px] left-[195px]">
-          {searchSchool?.map(school => (
-            <div key={school.schoolId} className="w-[466px] h-[58px] rounded-[5px] bg-white">
-              <span
-                onClick={handleClick}
-                id={school.schoolId}
-                className="flex text-[16px] pt-[20px] ml-[20px] items-center font-medium"
-              >
-                {school.schoolName}
-              </span>
-            </div>
-          ))}
-        </div>
+        {searchSchool?.map(school => (
+          <div
+            key={school.schoolId}
+            className="absolute top-[70px] left-[195px] w-[466px] h-[58px] rounded-[5px] bg-white"
+          >
+            <span
+              onClick={handleClick}
+              id={school.schoolId}
+              className="flex text-[16px] mt-[15px] ml-[20px] items-center font-medium"
+            >
+              {school.schoolName}
+            </span>
+          </div>
+        ))}
       </div>
       <div className="flex items-center mr-[32px] m-5">
         <div className="flex justify-center items-center w-[100px] cursor-pointer">
@@ -167,19 +195,25 @@ const RestaurantNavbar = () => {
         </div>
         <span className="w-[1px] h-[30.5px] bg-[#B1B1B1] mr-[19px]"></span>
         <div
-          onClick={handleClickLogout}
+          onClick={handleLogin}
           className="flex justify-center items-center w-[110px] cursor-pointer"
         >
-          <img src={Logout} alt="" className="mr-[6px] w-[16px] h-[16px] " />
-          <span className="text-[16px] fw-400 leading-[19px] text-[#353535] mr-[30px]">LOGOUT</span>
+          <img src={Logout} alt="" className="mr-[6px] w-[16px] h-[16px]" />
+          {isLogin ? (
+            <span className="text-[16px] fw-400 leading-[19px] text-[#353535] mr-[30px]">
+              LOGOUT
+            </span>
+          ) : (
+            <span className="text-[16px] fw-400 leading-[19px] text-[#353535] mr-[30px]">
+              LOGIN
+            </span>
+          )}
         </div>
-        <div className="flex flex-col items-center">
+        <div className="flex flex-col items-center cursor-pointer">
           <img
             src={`/image/${IMAGE_FILE_ID}`}
             className='relative top-[34px] bg-cover w-[40px] h-[40px] bg-[url("./img/circle_human.png")] rounded-full mb-[70px]'
-            onClick={() => {
-              history.push('/mypage');
-            }}
+            onClick={handleMypage}
           />
         </div>
       </div>
